+++ src/OptiWave/dct/__init__.py
@@ -1,0 +1,15 @@
+from .base_dct import (
+    normalize_image,
+    apply_dct,
+    apply_idct,
+    reconstruct_image,
+    dct_pipeline,
+)
+
+__all__ = [
+    "normalize_image",
+    "apply_dct",
+    "apply_idct",
+    "reconstruct_image",
+    "dct_pipeline",
+]
