+++ src/OptiWave/svd/block_svd.py
