+++ src/OptiWave/utils/image_loader.py
