--- src/dct/__init__.py
@@ -1,15 +1,0 @@
-from .base_dct import (
-    normalize_image,
-    apply_dct,
-    apply_idct,
-    reconstruct_image,
-    dct_pipeline,
-)
-
-__all__ = [
-    "normalize_image",
-    "apply_dct",
-    "apply_idct",
-    "reconstruct_image",
-    "dct_pipeline",
-]
